Extracted the client building to a factory.
